+++ CHANGELOG.md
@@ -5,4 +5,5 @@ FEATURES:
 - Created provider.
 - Created resources:
   - `fauna_collection` (Collection)
+  - `fauna_database` (Database)
   - `fauna_index` (Index)